+++ R/plot.R
@@ -36,12 +36,14 @@
 #' @importFrom stats binom.test p.adjust
 #'
 #' @return An \strong{invisible} list whose first element is a logic value
-#'   indicating whether the plot is successful. For \strong{SBS96Catalog}, the
-#'   list will have a second element, which is a numeric vector giving the
-#'   coordinates of all the bar midpoints drawn, useful for adding to the graph.
-#'   For \strong{SBS192Catalog} with "counts" catalog.type and non-NULL
-#'   abundance and \code{plot.SBS12 = TRUE}, the list will have a second element
-#'   which is a list containing the strand bias statistics.
+#'   indicating whether the plot is successful. For \code{SBS96Catalog},
+#'   \code{SBS192Catalog}, \code{DBS78Catalog}, \code{DBS144Catalog} and
+#'   \code{IndelCatalog}, the list will have a second element, which is a
+#'   numeric vector giving the coordinates of all the bar midpoints drawn,
+#'   useful for adding to the graph. For \strong{SBS192Catalog} with "counts"
+#'   catalog.type and non-NULL abundance and \code{plot.SBS12 = TRUE}, the list
+#'   will have an additional element which is a list containing the strand bias
+#'   statistics.
 #'   
 #' @note The sizes of repeats involved in deletions range from 0 to 5+ in the
 #'   mutational-spectra and signature catalog rownames, but for plotting and
@@ -604,9 +606,10 @@ PlotCatalog.SBS192Catalog <-
     
     # Check whether it is possible to return the p-values from binomial test
     if (isTRUE(plot.SBS12) && IsBinomialTestApplicable(catalog)) {
-      invisible(list(plot.success = TRUE, strand.bias.statistics = list0))
+      invisible(list(plot.success = TRUE, plot.object = bp, 
+                     strand.bias.statistics = list0))
     } else {
-      invisible(list(plot.success = TRUE))
+      invisible(list(plot.success = TRUE, plot.object = bp))
     }
 }
 
@@ -942,7 +945,7 @@ PlotCatalog.DBS78Catalog <- function(catalog, plot.SBS12, cex,
   text(bp, -ymax / 15, labels = substr(rownames(catalog), 3, 3),
        cex = 0.5, srt = 90, adj = 1, xpd = NA)
 
-  invisible(list(plot.success = TRUE))
+  invisible(list(plot.success = TRUE, plot.object = bp))
 }
 
 #' @export
@@ -1078,7 +1081,7 @@ PlotCatalog.DBS144Catalog <- function(catalog, plot.SBS12, cex = par("cex"),
   text(bp[8], ymax, labels = colnames(catalog), xpd = NA,
        font = 2, cex = cex, adj = c(0, 0))
 
-  invisible(list(plot.success = TRUE))
+  invisible(list(plot.success = TRUE, plot.object = bp))
 }
 
 #' @export
@@ -1555,7 +1558,7 @@ PlotCatalog.IndelCatalog <- function(catalog, plot.SBS12, cex,
   text(bp, -ymax * 0.15, labels = mut.type, cex = 0.65, xpd = NA)
   text(bottom.pos, -ymax * 0.27, labels = bottom.lab, cex = 0.75, xpd = NA)
 
-  invisible(list(plot.success = TRUE))
+  invisible(list(plot.success = TRUE, plot.object = bp))
 }
 
 #' @export