+++ NEWS.md
@@ -21,6 +21,10 @@ and `region` attribute as the source catalog, otherwise the function will raise 
 * Added an extra argument `suppress.discarded.variants.warnings` in *exported* functions
 `ReadAndSplitMutectVCFs`, `ReadAndSplitStrelkaSBSVCFs`, `VCFsToSBSCatalogs`,`VCFsToDBSCatalogs`,  `VCFsToIDCatalogs`, `MutectVCFFilesToCatalog`, `MutectVCFFilesToCatalogAndPlotToPdf`, `MutectVCFFilesToZipFile`.
 * Added legend for SBS192 plot and changed the legend text for SBS12 plot.
+* Added a second element `plot.object` to the returned list from function
+`PlotCatalog` for SBS192Catalog, DBS78Catalog, DBS144Catalog and IndelCatalog.
+The second element is a numeric vector giving the coordinates of all the bar
+midpoints drawn, useful for adding to the graph.
 
 ## Changed
 * **(Non backward-compatible)** Changed the return value of *exported* function
@@ -34,6 +38,8 @@ FALSE.
 * Minor changes to documentation of functions `PlotCatalog`, `PlotCatalogToPdf`, `StrelkaSBSVCFFilesToZipFile`, `StrelkaIDVCFFilesToZipFile` and `MutectVCFFilesToZipFile`.
 * Updated documentation for the return value of functions `StrelkaIDVCFFilesToCatalog`, `StrelkaIDVCFFilesToCatalogAndPlotToPdf`, `StrelkaIDVCFFilesToZipFile` and `VCFsToIDCatalogs` to make it clearer to the user.
 * Changed the return of functions `PlotCatalog` and `PlotCatalogToPdf` to an **invisible** list.
+* Optimized code in *exported* function `GetMutectVAF` and internal function `CanonicalizeDBS`,
+`CanonicalizeQUAD` to make them run faster.
 
 ## Fixed
 * Fixed bugs in `if` statement in internal functions